+++ demos/tutorial/src/main/scala/com/sfxcode/sapphire/core/demo/tutorial/ApplicationController.scala
@@ -2,7 +2,6 @@ package com.sfxcode.sapphire.core.demo.tutorial
 
 import javax.enterprise.context.ApplicationScoped
 import javax.inject.Named
-
 import com.sfxcode.sapphire.core.controller.AppController
 import com.sfxcode.sapphire.core.demo.tutorial.controller.MainWindowController
 
@@ -10,11 +9,23 @@ import com.sfxcode.sapphire.core.demo.tutorial.controller.MainWindowController
 @ApplicationScoped
 class ApplicationController extends AppController {
 
-  lazy val mainWindowController = getController[MainWindowController]()
+  lazy val mainWindowController: MainWindowController =
+    getController[MainWindowController]()
 
   def applicationDidLaunch() {
     logger.info("start " + this)
     applicationEnvironment.loadResourceBundle("bundles/application")
     replaceSceneContent(mainWindowController)
+  }
+
+  def reload(): Unit = {
+    // Styling
+    reloadStyles()
+    // Resources
+    applicationEnvironment.clearResourceBundleCache()
+    applicationEnvironment.loadResourceBundle("bundles/application")
+    // FXML
+    val newMainWindowController = getController[MainWindowController]()
+    replaceSceneContent(newMainWindowController)
   }
 }