+++ demos/tutorial/src/main/scala/com/sfxcode/sapphire/core/demo/tutorial/controller/AbstractViewController.scala
@@ -2,6 +2,7 @@
 package com.sfxcode.sapphire.core.demo.tutorial.controller
 
 import com.sfxcode.sapphire.core.controller.ViewController
+import com.sfxcode.sapphire.core.scene.ContentManager
 import com.typesafe.scalalogging.LazyLogging
 
 abstract class AbstractViewController extends ViewController with LazyLogging {
@@ -14,16 +15,9 @@ abstract class AbstractViewController extends ViewController with LazyLogging {
     parent.asInstanceOf[MainWindowController]
   }
 
-  /**
-   *
-   * @return
-   */
-  def statusBarController = getBean[StatusBarController]()
+  def statusBarController: StatusBarController = getBean[StatusBarController]()
 
-  /**
-   *
-   * @return workspace manager resolved by parent
-   */
-  def workspaceManager = mainWindowController.workspaceManager
+
+  def workspaceManager: ContentManager = mainWindowController.workspaceManager
 
 }