+++ demos/tutorial/src/main/scala/com/sfxcode/sapphire/core/demo/tutorial/controller/PersonController.scala
@@ -4,14 +4,14 @@ import javafx.event.ActionEvent
 import javafx.fxml.FXML
 import javafx.scene.control.TableView
 import javafx.scene.layout.GridPane
-
 import com.sfxcode.sapphire.core.Includes._
-import com.sfxcode.sapphire.core.demo.tutorial.model.{ Person, PersonFactory }
-import com.sfxcode.sapphire.core.value.{ KeyBindings, FXBeanAdapter, FXBean }
-
+import com.sfxcode.sapphire.core.demo.tutorial.model.{Person, PersonFactory}
+import com.sfxcode.sapphire.core.fxml.FxmlLoader
+import com.sfxcode.sapphire.core.value.{FXBean, FXBeanAdapter, KeyBindings}
 import scalafx.Includes._
 import scalafx.collections.ObservableBuffer
 
+@FxmlLoader(path="/fxml/widget/Person.fxml")
 class PersonController extends AbstractViewController {
 
   @FXML
@@ -20,7 +20,7 @@ class PersonController extends AbstractViewController {
   @FXML
   var editPane: GridPane = _
 
-  lazy val adapter = FXBeanAdapter[Person](this)
+  lazy val adapter: FXBeanAdapter[Person] = FXBeanAdapter[Person](this)
 
   def items: ObservableBuffer[FXBean[Person]] = PersonFactory.personList