+++ demos/tutorial/src/main/scala/com/sfxcode/sapphire/core/demo/tutorial/controller/SecondWorkspaceController.scala
@@ -1,8 +1,0 @@
-package com.sfxcode.sapphire.core.demo.tutorial.controller
-
-/**
- * Created by tom on 20.10.15.
- */
-class SecondWorkspaceController extends AbstractViewController {
-
-}