+++ demos/tutorial/src/main/scala/com/sfxcode/sapphire/core/demo/tutorial/controller/StatusBarController.scala
@@ -1,32 +1,31 @@
 package com.sfxcode.sapphire.core.demo.tutorial.controller
 
-import javax.enterprise.context.ApplicationScoped
-import javax.inject.Named
-
 import com.sfxcode.sapphire.core.controller.ViewController
 import com.typesafe.scalalogging.LazyLogging
-
+import javax.enterprise.context.ApplicationScoped
+import javax.inject.Named
 import scalafx.Includes._
 import scalafx.event.ActionEvent
 import scalafx.geometry.Insets
-import scalafx.scene.control.{ Button, Label }
+import scalafx.scene.control.{Button, Label}
 import scalafx.scene.layout.HBox
 
 @Named
 @ApplicationScoped
 class StatusBarController extends ViewController with LazyLogging {
 
-  val actionLabel = new Label {
+  val actionLabel: Label = new Label {
     text = "Status Bar Action Label ..."
     padding = Insets(5)
   }
 
-  val statusLabel = new Label {
+  val statusLabel: Label = new Label {
     text = "Status Bar Status Label ..."
     padding = Insets(5)
   }
 
   rootPane = new HBox {
+    id = "statusBar"
     padding = Insets(5)
     spacing = 10
     children = List(
@@ -42,7 +41,7 @@ class StatusBarController extends ViewController with LazyLogging {
       actionLabel)
   }
 
-  def handleButtonAction(button: Button) =
+  def handleButtonAction(button: Button): ActionEvent => Unit =
     (_: ActionEvent) => {
       logger.debug("%s".format(button.text))
       updateLabel(button)