+++ demos/tutorial/src/main/scala/com/sfxcode/sapphire/core/demo/tutorial/model/Person.scala
@@ -5,5 +5,5 @@ case class Person(id: Int, name: String, age: Int)
 object PersonFactory {
   val personList = List(Person(1, "Tom", 42), Person(2, "Tim", 43), Person(3, "Bob", 44))
 
-  def person = personList.head
+  def person: Person = personList.head
 }