+++ src/it/scala/com/sfxcode/sapphire/core/test/integration/controller/DisplayInfoController.scala
@@ -1,9 +1,0 @@
-package com.sfxcode.sapphire.core.test.integration.controller
-
-import com.sfxcode.sapphire.core.controller.ViewController
-
-class DisplayInfoController extends ViewController {
-  override def didGainVisibility(): Unit = {
-    super.didGainVisibility()
-  }
-}