+++ src/it/scala/com/sfxcode/sapphire/core/test/integration/controller/SecondWorkspaceController.scala
@@ -1,7 +1,0 @@
-package com.sfxcode.sapphire.core.test.integration.controller
-
-class SecondWorkspaceController extends AbstractWorkspaceController {
-
-  val testString = "Second Workspace Controller Test String"
-
-}