+++ src/it/scala/com/sfxcode/sapphire/core/test/integration/model/Person.scala
@@ -1,7 +1,0 @@
-package com.sfxcode.sapphire.core.test.integration.model
-
-case class Person(id: Int, name: String, age: Int)
-
-object PersonFactory {
-  val personList = List(Person(1, "Tom", 42), Person(2, "Tim", 43), Person(3, "Bob", 44))
-}