+++ src/main/scala/com/sfxcode/sapphire/core/cdi/FXApp.scala
@@ -1,29 +1,26 @@
 package com.sfxcode.sapphire.core.cdi
 
+import com.sfxcode.sapphire.core.base.ConfigValues
 import javax.enterprise.util.AnnotationLiteral
-
 import com.sfxcode.sapphire.core.cdi.annotation._
 import com.sfxcode.sapphire.core.cdi.provider.ApplicationParametersProvider
-import com.typesafe.config.ConfigFactory
 import org.apache.deltaspike.core.api.provider.BeanProvider
-
 import scalafx.application.JFXApp
 import scalafx.application.JFXApp.PrimaryStage
 import scalafx.scene.Scene
 import scalafx.scene.shape.Rectangle
 import scalafx.stage.Stage
 
-abstract class FXApp extends JFXApp {
-  val configuration = ConfigFactory.load()
+abstract class FXApp extends JFXApp with ConfigValues{
 
   init(applicationStage)
 
   def applicationStage: Stage = createPrimaryStage()
 
   def createPrimaryStage(
-    stageWidth: Int = configuration.getInt("sapphire.core.defaultStage.width"),
-    stageHeight: Int = configuration.getInt("sapphire.core.defaultStage.height"),
-    stageTitle: String = configuration.getString("sapphire.core.defaultStage.title")): PrimaryStage = new PrimaryStage {
+    stageWidth: Int = configIntValue("sapphire.core.defaultStage.width"),
+    stageHeight: Int = configIntValue("sapphire.core.defaultStage.height"),
+    stageTitle: String = configStringValue("sapphire.core.defaultStage.title")): PrimaryStage = new PrimaryStage {
     title = stageTitle
     width = stageWidth
     height = stageHeight