+++ src/main/scala/com/sfxcode/sapphire/core/controller/AppController.scala
@@ -3,13 +3,14 @@ package com.sfxcode.sapphire.core.controller
 import java.util.ResourceBundle
 
 import javax.enterprise.event.Observes
-import com.sfxcode.sapphire.core.cdi.annotation.{ FXStage, Startup }
+import com.sfxcode.sapphire.core.cdi.annotation.{FXStage, Startup}
 import com.sfxcode.sapphire.core.fxml.FxmlLoading
 import com.sfxcode.sapphire.core.scene.NodeLocator
+import com.sun.javafx.css.StyleManager
 import com.typesafe.scalalogging.LazyLogging
 import scalafx.application.Platform
 import scalafx.collections.ObservableMap
-import scalafx.scene.{ Parent, Scene }
+import scalafx.scene.{Parent, Scene}
 import scalafx.stage.Stage
 
 abstract class AppController extends FxmlLoading with NodeLocator with LazyLogging {
@@ -85,6 +86,10 @@ abstract class AppController extends FxmlLoading with NodeLocator with LazyLoggi
 
   def applicationWillStop(): Unit = {
     logger.debug("exit in Progress")
+  }
+
+  def reloadStyles(): Unit = {
+    StyleManager.getInstance().stylesheetContainerMap.clear()
   }
 
   def exit(): Unit = {