+++ src/main/scala/com/sfxcode/sapphire/core/fxml/FxmlLoader.scala
@@ -1,0 +1,25 @@
+package com.sfxcode.sapphire.core.fxml
+import scala.reflect.runtime.{universe => ru}
+import ru._
+
+case class FxmlLoader(path: String) extends scala.annotation.StaticAnnotation
+
+object FxmlLoader  {
+
+  def pathValue(clazzTag: scala.reflect.ClassTag[_]): String = {
+    var result = ""
+    val runtimeClass = clazzTag.runtimeClass
+    val rootMirror = ru.runtimeMirror(runtimeClass.getClassLoader)
+    val myAnnotatedClass = rootMirror.classSymbol(runtimeClass)
+    val annotation: Option[ru.Annotation] = myAnnotatedClass.annotations.find(_.tree.tpe =:= ru.typeOf[FxmlLoader])
+    annotation.flatMap { a =>
+      a.tree.children.tail.collectFirst {
+        case Literal(Constant(name: String)) =>
+          result = name
+      }
+    }
+    result
+  }
+
+
+}